% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{edit_aws_creds}
\alias{edit_aws_creds}
\alias{aws_creds}
\alias{aws_creds_path}
\title{Edit your AWS credentials}
\usage{
edit_aws_creds()

aws_creds(file = aws_creds_path())

aws_creds_path()
}
\description{
Edit your AWS credentials
}
