% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_get_permissions}
\alias{aws_get_permissions}
\title{Lookup user AWS information and permissions}
\usage{
aws_get_permissions(
  username = NULL,
  actions = list("s3:CreateBucket", "s3:ListBuckets", "iam:CreateUser",
    "iam:CreateAccessKey"),
  credentials = aws_creds()$default
)
}
\arguments{
\item{username}{The AWS account username, if NULL, defaults to that
associated with current credentials}

\item{actions}{A list of permissions to check for}

\item{credentials}{A list with members \code{aws_access_key_id} and
\code{aws_secret_access_key}. Defaults to lookup up the user \code{.aws/credentials}
rather than environment variables.}
}
\description{
Lookup user AWS information and permissions
}
