% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{ehaproj_config}
\alias{ehaproj_config_path}
\alias{edit_ehaproj_config}
\title{Project configuration}
\usage{
ehaproj_config(..., .options = list(), defaults = FALSE)

ehaproj_config_path()

edit_ehaproj_config()
}
\arguments{
\item{...}{name-value pairs of configuration options to set}

\item{.options}{name-value pairs of configuration options, as list}

\item{defaults}{Return package defaults, unmodified by any user config,
options, or variables}
}
\description{
\code{ehaproj} lets you set preferences for default new project configuration in
several different ways:
}
\details{
\enumerate{
\item If an environment variable of the form \verb{EHAPROJ_CONFIG_<NAME>} is available,
then that option's value will be used;
\item If an \R option of the form \verb{ehaproj.<name>} is available,
then that option's value will be used;
\item If the option is set in the user configuration YAML file, found at
\code{ehaproj_config_path()}. This file is modified if the user opts to save
settings as defaults when creating a project. You can edit it with
\code{edit_ehaproj_config()};
\item Otherwise, the default for that particular configuration value is used.
You can see these defaults wiht \code{ehaproj_config(defaults = TRUE)}
}

For example, the
configuration option \code{parent_dir} could be configured as:
\itemize{
\item \verb{options(ehaproj.parent_dir = <...>)}
\item \verb{Sys.setenv(EHAPROJ_PARENT_DIR = <...>)}
\item \verb{parent_dir: <...>} in the configuration YAML file.
}

If configration options are prepended with \verb{ask:} (as in \code{use_git: "ask:TRUE"}),
the user will be prompted for an answer with the option being provided as
a default.  Several options can be set to \code{"i"}, in which case \code{ehaproj()}
will launch interactive dialogs for picking from options.
}
\examples{

# set your default directory for projects
options(ehaproj.parent_dir = "~/projects")

#' # set your default github namespace
Sys.setenv(EHAPROJ_GITHUB_ORG = "ecohealthalliance")

# always use git-crypt in projects
Sys.setenv(EHAPROJ_USE_CRYPT = TRUE)

}
