% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehaproj-package.R
\docType{package}
\name{ehaproj-package}
\alias{ehaproj}
\alias{ehaproj-package}
\title{ehaproj: Provides tools for setting up and running projects using EcoHealth Alliance templates and workflows}
\description{
Provides functions for uploading and emailing documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliance/ehaproj}
  \item Report bugs at \url{https://github.com/ecohealthalliance/ehaproj/issues}
}

}
\author{
\strong{Maintainer}: Collin Schwantes \email{schwantes@ecohealthalliance.org} (\href{https://orcid.org/0000-0002-9882-941X}{ORCID})

Authors:
\itemize{
  \item Noam Ross \email{ross@ecohealthalliance.org} (\href{https://orcid.org/0000-0002-2136-0000}{ORCID})
}

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder]
}

}
\keyword{internal}
