% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ehaproj}
\alias{ehaproj}
\title{Create a new project}
\usage{
ehaproj(..., .options = list())
}
\arguments{
\item{...}{project configuration options.  See \code{ehaproj_config()} for all
options}

\item{.options}{configuration options passed as a list}
}
\description{
Create a new project
}
