% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_files}
\alias{env_files}
\title{List of files starting with \code{.env}}
\usage{
env_files(path = ".", regexp = "^\\\\.env[^/]*$", ...)
}
\arguments{
\item{path}{directory to search in}

\item{regexp}{matching regular expression}

\item{...}{other arguments to \code{\link[fs:dir_ls]{fs::dir_ls()}}}
}
\description{
List of files starting with \code{.env}
}
