% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_projects.R
\name{gh_create_project}
\alias{gh_create_project}
\title{Create a GitHub Project Board}
\usage{
gh_create_project(organization = NULL, template = NULL, repo = NULL)
}
\arguments{
\item{organization}{or user namespace. NULL uses the current user}

\item{template}{a GitHub Project to use as a template. Can take name, number or GraphQL ID of a project, preceeded with "organiation/"}

\item{repo}{A GitHub repository to associate the project with}
}
\description{
This creates a GitHub project.  Defaults are to use EHA projects templates and automations
}
