% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_list_repos}
\alias{gh_list_repos}
\title{Get a list of repositories under a GitHub namespace,}
\usage{
gh_list_repos(org = NULL)
}
\arguments{
\item{org}{The GitHub organization. If NULL, defaults to the user's space}
}
\description{
Used to test if a project name already exists
}
