% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitcrypt.R
\name{gitcrypt_unlock}
\alias{gitcrypt_unlock}
\alias{gitcrypt_lock}
\alias{is_file_gitencrypted}
\title{Checks if a file is encrypted with git-crypt.}
\usage{
gitcrypt_unlock(dir = ".")

gitcrypt_lock(dir = ".", force = FALSE, all = FALSE, key = NULL)

is_file_gitencrypted(f)
}
\arguments{
\item{dir}{the directory to operate on}

\item{force}{lock even though the repository is not clean}

\item{all}{lock all keys, not just default}

\item{key}{lock a specific key}

\item{f}{a file to check}
}
\description{
Failing/forgetting to run "\verb{git-crypt unlock} when checking out an encrypted
repository is a common source of errors.  Such files start with the header
\verb{^@GITCRYPT^@} (where \verb{^@} are null binary bytes), which this function
checks for.
}
