% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_projects.R
\name{gql}
\alias{gql}
\title{Same as gh::gh_gql, but use updated Global IDs in header by default}
\usage{
gql(query, ...)
}
\description{
Same as gh::gh_gql, but use updated Global IDs in header by default
}
