% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{load_dotenv}
\alias{load_dotenv}
\title{Like dotenv::load_dot_env but allows external setting to override if desired}
\usage{
load_dotenv(
  files = env_files(),
  override = c("exported", "all", "none"),
  warning_on_error = TRUE
)
}
\arguments{
\item{files}{Files to load environment variables from, The default, \code{\link[=env_files]{env_files()}},
loads all files in the working directory starting with \verb{.env*}, in alphabetical
order}

\item{override}{Whether variables in \verb{.env*} should override already-set
environment variables. The default, "exported" only does this for variables
prefixed with \code{export}.}
}
\description{
Also checks if file is binary and/or encrypted and reports a message back.
}
