% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile_options.R
\name{reconcile_options}
\alias{reconcile_options}
\title{Reconcile options with each other and the user environment}
\usage{
reconcile_options(options, user_sitrep = get_sitrep_user())
}
\arguments{
\item{options}{a list of ehaproj configuration options}

\item{sitrep}{a list produced by \code{\link[=get_sitrep_user]{get_sitrep_user()}}.}
}
\description{
Reconcile options with each other and the user environment
}
\author{
'Noam Ross'
}
