% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_in_paths_and_files}
\alias{rename_in_paths_and_files}
\alias{rename_in_filenames}
\alias{rename_in_file}
\title{Find and replace in files and filenames
Rename stuff}
\usage{
rename_in_paths_and_files(dir, patterns, replacements)

rename_in_filenames(rename_in_filenames, patterns, replacements)

rename_in_file(file, patterns, replacements)
}
\arguments{
\item{dir}{directory in which to perform the operation}

\item{patterns}{a vector of fixed strings to find and replace}

\item{replacements}{a vector of replacements for each of the fixed strings}

\item{file}{a file to be modified}

\item{filenames}{filenames that should be modified}
}
\description{
Find and replace in files and filenames
Rename stuff
}
