% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{get_sitrep}
\alias{get_sitrep}
\alias{get_sitrep_user}
\alias{get_sitrep_project}
\title{Situation Reports}
\usage{
get_sitrep()

get_sitrep_user()

get_sitrep_project()
}
\value{
A list of statuses, with NA values being indeterminate.  The list
is of class \code{ehaproj_status} and has a pretty-printing method.
}
\description{
These functions check the staus of machine configuration of or project
setup.
}
\examples{
  sitrep()
}
