% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitcrypt.R
\name{use_gitcrypt}
\alias{use_gitcrypt}
\title{whoami::email_address(), which defaults to the primary git user.}
\usage{
use_gitcrypt(dir = getwd(), gpg_user = NULL, verbose = TRUE)
}
\description{
whoami::email_address(), which defaults to the primary git user.
}
